Create new dir structure
