+++ tests/test_resid_mlp.py
@@ -51,7 +51,6 @@ def test_resid_mlp_decomposition_happy_path() -> None:
         n_random_masks=2,
         param_match_coeff=1.0,
         masked_recon_coeff=1,
-        act_recon_coeff=1,
         lp_sparsity_coeff=1.0,
         pnorm=0.9,
         lr=1e-3,
